test(apps/web): Add unit tests for GenericInput form
