+++ packages/ui/src/GenericInputForm/initialValues.ts
@@ -1,0 +1,16 @@
+import { FormValues } from "./types";
+
+export const initialValues: FormValues = {
+    mode: "hex",
+    application: "",
+    rawInput: "0x",
+    stringInput: "",
+    abiMethod: "existing",
+    specificationMode: "json_abi",
+    humanAbi: "",
+    abiParam: "",
+    savedAbiParam: "",
+    specificationId: "",
+    abiFunctionName: "",
+    abiFunctionParams: [],
+};
